--- a/news/cache-executables-in.rst
+++ b/news/cache-executables-in.rst
@@ -1,0 +1,23 @@
+**Added:**
+
+* Added cache on finding executable files, to speedup command-not-found suggestions
+
+**Changed:**
+
+* <news item>
+
+**Deprecated:**
+
+* <news item>
+
+**Removed:**
+
+* <news item>
+
+**Fixed:**
+
+* <news item>
+
+**Security:**
+
+* <news item>