This is a real shot in the dark! But sometimes it bugged me how when you put a not recognized command, it takes a second to respond the error. It looks like the vast majority of this time is finding suggestions. Specifically in executables, with has two for loops that go through files in the path. There's probably do a lot of redundancy in that.

If we put a simple cache on that function, it speeds up **dramatically** when there are no suggestions!

BEFORE:
```
$ timeit "asdfasdfasdf"
xonsh: subprocess mode: command not found: asdfasdfasdf
xonsh: subprocess mode: command not found: asdfasdfasdf
xonsh: subprocess mode: command not found: asdfasdfasdf
1 loops, best of 3: 1.77 s per loop
```

AFTER:
```
$ timeit "asdfasdfasdf"
...
xonsh: subprocess mode: command not found: asdfasdfasdf
xonsh: subprocess mode: command not found: asdfasdfasdf
xonsh: subprocess mode: command not found: asdfasdfasdf
10 loops, best of 3: 35 ms per loop
```

On the other hand, there's a very small (but basically inconsequential) speedup when it does find a suggestion.

BEFORE:
```
$ timeit "ca"
...
Did you mean one of the following?
    cd:  Alias
    fg:  Alias
    bg:  Alias
    ls:  Alias
10 loops, best of 3: 50.5 ms per loop
```

AFTER:
```
$ timeit "ca"
...
Did you mean one of the following?
    cd:  Alias
    fg:  Alias
    bg:  Alias
    ls:  Alias
10 loops, best of 3: 47.5 ms per loop
```