@scopatz 2020-09-18
Hey @dannysepler - thanks for putting this in. I definitely don't think the failure should take 2 sec to print the message.

However, I don't think that this is a safe fix. Folks can modify their environment between two commands (say by installing a pakage) and you would expect that after the install the new cmd would be available.  

I think a better fix would be for the alternatives message to use the lazy APIs on the CommandsCache.  The CommandsCache itself uses the `executable_in()` function.  Sorry to make this more difficult.