@dannysepler 2020-09-18
not at all! these test failures were making me second guess myself anyway. thanks for giving me that direction!