@scopatz 2020-09-18
This looks much better! What does the timing look like with this now?