@dannysepler 2020-09-18
thanks, @scopatz! i've been updating the timing as i update the code. I'm not sure what was going on with the 1.77 seconds in my first version of this diff. my laptop now says it's closer to 500 ms -> 90 ms. It's still a significant speedup, but not quite the insanity I saw before :)